Fix clippy warnings

+++ tests/sbert.rs
@@ -24,19 +24,19 @@ fn sbert_distilbert() -> anyhow::Result<()> {
     let sentences = ["This is an example sentence", "Each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - -0.03479306).abs() < 1e-4);
-    assert!((embeddings[0][1] - 0.02635195).abs() < 1e-4);
-    assert!((embeddings[0][2] - -0.04427199).abs() < 1e-4);
-    assert!((embeddings[0][509] - 0.01743882).abs() < 1e-4);
-    assert!((embeddings[0][510] - -0.01952395).abs() < 1e-4);
-    assert!((embeddings[0][511] - -0.00118101).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - -0.03479306).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - 0.02635195).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - -0.04427199).abs() < 1e-4);
+    assert!((embeddings[0][509] as f64 - 0.01743882).abs() < 1e-4);
+    assert!((embeddings[0][510] as f64 - -0.01952395).abs() < 1e-4);
+    assert!((embeddings[0][511] as f64 - -0.00118101).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - 0.02096637).abs() < 1e-4);
-    assert!((embeddings[1][1] - -0.00401743).abs() < 1e-4);
-    assert!((embeddings[1][2] - -0.05093712).abs() < 1e-4);
-    assert!((embeddings[1][509] - 0.03618195).abs() < 1e-4);
-    assert!((embeddings[1][510] - 0.0294408).abs() < 1e-4);
-    assert!((embeddings[1][511] - -0.04497765).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - 0.02096637).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - -0.00401743).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - -0.05093712).abs() < 1e-4);
+    assert!((embeddings[1][509] as f64 - 0.03618195).abs() < 1e-4);
+    assert!((embeddings[1][510] as f64 - 0.0294408).abs() < 1e-4);
+    assert!((embeddings[1][511] as f64 - -0.04497765).abs() < 1e-4);
 
     Ok(())
 }
@@ -63,19 +63,19 @@ fn sbert_bert() -> anyhow::Result<()> {
     let sentences = ["this is an example sentence", "each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - -0.153149).abs() < 1e-4);
-    assert!((embeddings[0][1] - 0.12944254).abs() < 1e-4);
-    assert!((embeddings[0][2] - -0.03903132).abs() < 1e-4);
-    assert!((embeddings[0][765] - 0.09051804).abs() < 1e-4);
-    assert!((embeddings[0][766] - -0.12790504).abs() < 1e-4);
-    assert!((embeddings[0][767] - -0.02097229).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - -0.153149).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - 0.12944254).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - -0.03903132).abs() < 1e-4);
+    assert!((embeddings[0][765] as f64 - 0.09051804).abs() < 1e-4);
+    assert!((embeddings[0][766] as f64 - -0.12790504).abs() < 1e-4);
+    assert!((embeddings[0][767] as f64 - -0.02097229).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - -0.01726065).abs() < 1e-4);
-    assert!((embeddings[1][1] - 0.17196466).abs() < 1e-4);
-    assert!((embeddings[1][2] - 0.14468355).abs() < 1e-4);
-    assert!((embeddings[1][765] - 0.26710105).abs() < 1e-4);
-    assert!((embeddings[1][766] - 0.15404876).abs() < 1e-4);
-    assert!((embeddings[1][767] - -0.14560737).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - -0.01726065).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - 0.17196466).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - 0.14468355).abs() < 1e-4);
+    assert!((embeddings[1][765] as f64 - 0.26710105).abs() < 1e-4);
+    assert!((embeddings[1][766] as f64 - 0.15404876).abs() < 1e-4);
+    assert!((embeddings[1][767] as f64 - -0.14560737).abs() < 1e-4);
 
     Ok(())
 }
@@ -102,19 +102,19 @@ fn sbert_bert_small() -> anyhow::Result<()> {
     let sentences = ["this is an example sentence", "each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - -2.02682902e-04).abs() < 1e-4);
-    assert!((embeddings[0][1] - 8.14802647e-02).abs() < 1e-4);
-    assert!((embeddings[0][2] - 3.13617811e-02).abs() < 1e-4);
-    assert!((embeddings[0][381] - 6.20930083e-02).abs() < 1e-4);
-    assert!((embeddings[0][382] - 4.91031967e-02).abs() < 1e-4);
-    assert!((embeddings[0][383] - -2.90199649e-04).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - -2.02682902e-04).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - 8.14802647e-02).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - 3.13617811e-02).abs() < 1e-4);
+    assert!((embeddings[0][381] as f64 - 6.20930083e-02).abs() < 1e-4);
+    assert!((embeddings[0][382] as f64 - 4.91031967e-02).abs() < 1e-4);
+    assert!((embeddings[0][383] as f64 - -2.90199649e-04).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - 6.47571534e-02).abs() < 1e-4);
-    assert!((embeddings[1][1] - 4.85198125e-02).abs() < 1e-4);
-    assert!((embeddings[1][2] - -1.78603437e-02).abs() < 1e-4);
-    assert!((embeddings[1][381] - 3.37569155e-02).abs() < 1e-4);
-    assert!((embeddings[1][382] - 8.43371451e-03).abs() < 1e-4);
-    assert!((embeddings[1][383] - -6.00359812e-02).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - 6.47571534e-02).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - 4.85198125e-02).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - -1.78603437e-02).abs() < 1e-4);
+    assert!((embeddings[1][381] as f64 - 3.37569155e-02).abs() < 1e-4);
+    assert!((embeddings[1][382] as f64 - 8.43371451e-03).abs() < 1e-4);
+    assert!((embeddings[1][383] as f64 - -6.00359812e-02).abs() < 1e-4);
 
     Ok(())
 }
@@ -139,19 +139,19 @@ fn sbert_distilroberta() -> anyhow::Result<()> {
     let sentences = ["This is an example sentence", "Each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - -0.03375624).abs() < 1e-4);
-    assert!((embeddings[0][1] - -0.06316338).abs() < 1e-4);
-    assert!((embeddings[0][2] - -0.0316612).abs() < 1e-4);
-    assert!((embeddings[0][765] - 0.03684864).abs() < 1e-4);
-    assert!((embeddings[0][766] - -0.02036646).abs() < 1e-4);
-    assert!((embeddings[0][767] - -0.01574).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - -0.03375624).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - -0.06316338).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - -0.0316612).abs() < 1e-4);
+    assert!((embeddings[0][765] as f64 - 0.03684864).abs() < 1e-4);
+    assert!((embeddings[0][766] as f64 - -0.02036646).abs() < 1e-4);
+    assert!((embeddings[0][767] as f64 - -0.01574).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - -0.01409588).abs() < 1e-4);
-    assert!((embeddings[1][1] - 0.00091114).abs() < 1e-4);
-    assert!((embeddings[1][2] - -0.00096315).abs() < 1e-4);
-    assert!((embeddings[1][765] - -0.02571585).abs() < 1e-4);
-    assert!((embeddings[1][766] - -0.00289072).abs() < 1e-4);
-    assert!((embeddings[1][767] - -0.00579975).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - -0.01409588).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - 0.00091114).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - -0.00096315).abs() < 1e-4);
+    assert!((embeddings[1][765] as f64 - -0.02571585).abs() < 1e-4);
+    assert!((embeddings[1][766] as f64 - -0.00289072).abs() < 1e-4);
+    assert!((embeddings[1][767] as f64 - -0.00579975).abs() < 1e-4);
 
     Ok(())
 }
@@ -178,19 +178,19 @@ fn sbert_albert() -> anyhow::Result<()> {
     let sentences = ["this is an example sentence", "each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - 0.20412037).abs() < 1e-4);
-    assert!((embeddings[0][1] - 0.48823047).abs() < 1e-4);
-    assert!((embeddings[0][2] - 0.5664698).abs() < 1e-4);
-    assert!((embeddings[0][765] - -0.37474486).abs() < 1e-4);
-    assert!((embeddings[0][766] - 0.0254627).abs() < 1e-4);
-    assert!((embeddings[0][767] - -0.6846024).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - 0.20412037).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - 0.48823047).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - 0.5664698).abs() < 1e-4);
+    assert!((embeddings[0][765] as f64 - -0.37474486).abs() < 1e-4);
+    assert!((embeddings[0][766] as f64 - 0.0254627).abs() < 1e-4);
+    assert!((embeddings[0][767] as f64 - -0.6846024).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - 0.25720373).abs() < 1e-4);
-    assert!((embeddings[1][1] - 0.24648172).abs() < 1e-4);
-    assert!((embeddings[1][2] - -0.2521183).abs() < 1e-4);
-    assert!((embeddings[1][765] - 0.4667896).abs() < 1e-4);
-    assert!((embeddings[1][766] - 0.14219822).abs() < 1e-4);
-    assert!((embeddings[1][767] - 0.3986863).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - 0.25720373).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - 0.24648172).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - -0.2521183).abs() < 1e-4);
+    assert!((embeddings[1][765] as f64 - 0.4667896).abs() < 1e-4);
+    assert!((embeddings[1][766] as f64 - 0.14219822).abs() < 1e-4);
+    assert!((embeddings[1][767] as f64 - 0.3986863).abs() < 1e-4);
 
     Ok(())
 }
@@ -216,19 +216,19 @@ fn sbert_t5() -> anyhow::Result<()> {
     let sentences = ["This is an example sentence", "Each sentence is converted"];
     let embeddings = sbert.encode(&sentences)?;
 
-    assert!((embeddings[0][0] - -0.00904849).abs() < 1e-4);
-    assert!((embeddings[0][1] - 0.0191336).abs() < 1e-4);
-    assert!((embeddings[0][2] - 0.02657794).abs() < 1e-4);
-    assert!((embeddings[0][765] - -0.00876413).abs() < 1e-4);
-    assert!((embeddings[0][766] - -0.05602207).abs() < 1e-4);
-    assert!((embeddings[0][767] - -0.02163094).abs() < 1e-4);
+    assert!((embeddings[0][0] as f64 - -0.00904849).abs() < 1e-4);
+    assert!((embeddings[0][1] as f64 - 0.0191336).abs() < 1e-4);
+    assert!((embeddings[0][2] as f64 - 0.02657794).abs() < 1e-4);
+    assert!((embeddings[0][765] as f64 - -0.00876413).abs() < 1e-4);
+    assert!((embeddings[0][766] as f64 - -0.05602207).abs() < 1e-4);
+    assert!((embeddings[0][767] as f64 - -0.02163094).abs() < 1e-4);
 
-    assert!((embeddings[1][0] - -0.00785422).abs() < 1e-4);
-    assert!((embeddings[1][1] - 0.03018173).abs() < 1e-4);
-    assert!((embeddings[1][2] - 0.03129675).abs() < 1e-4);
-    assert!((embeddings[1][765] - -0.01246878).abs() < 1e-4);
-    assert!((embeddings[1][766] - -0.06240674).abs() < 1e-4);
-    assert!((embeddings[1][767] - -0.00590969).abs() < 1e-4);
+    assert!((embeddings[1][0] as f64 - -0.00785422).abs() < 1e-4);
+    assert!((embeddings[1][1] as f64 - 0.03018173).abs() < 1e-4);
+    assert!((embeddings[1][2] as f64 - 0.03129675).abs() < 1e-4);
+    assert!((embeddings[1][765] as f64 - -0.01246878).abs() < 1e-4);
+    assert!((embeddings[1][766] as f64 - -0.06240674).abs() < 1e-4);
+    assert!((embeddings[1][767] as f64 - -0.00590969).abs() < 1e-4);
 
     Ok(())
 }
